[EventPipe] Restrict SequencePoint sequence numbers
Previously, EventPipeThread's lifetimes were inherently tied to the
lifetime of all SequencePoints that track the thread's last sequence
number. In adding ThreadSessionState cleanup, SequencePoints forced
deletions to be deferred.

SequencePoints do not need to extend EventPipeThreads lifetimes, as they
only need to track the last known sequence number of that thread by the
time the SequencePoint is written. To disassociate the two lifetimes and
allow ThreadSessionStates to be cleaned up as soon as possible:

1) Update SequencePoints in tandem with event processing
Delaying SequencePoint updating until the SequencePoint is written
requires the buffer_manager's thread_session_state_list to not trim any
ThreadSessionStates. Updating the mapped ThreadId:sequence_number in
tandem allows the thread_session_state_list to be trimmed immediately.

2) Restrict ThreadId:sequence_number tracking
Previously, SequencePoints tracked all threads and their last known
sequence number regardless of whether there have been new events, and
mandated reactionary removal when the ThreadSessionState was
being deleted. Instead, by tightening the conditions for when a
SequencePoint should track a particulr thread, they should never need to
track threads unnecessarily. The conditions for when a SequencePoint
needs to track a thread are:
  a) EventPipeThread is registered during SequencePoint init
  or
  b) EventPipeThread is unregistered during SP init and there are new
  event write attempts since the last SP